requirements.txt
anthropic>=0.43.0
playwright>=1.44.0
google-cloud-firestore>=2.16.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
requests>=2.31.0
